GG-25278: change timeout to int, fix JDBC

+++ modules/indexing/src/test/java/org/apache/ignite/internal/processors/query/timeout/DefaultQueryTimeoutThinJavaTest.java
@@ -38,9 +38,9 @@ public class DefaultQueryTimeoutThinJavaTest extends AbstractDefaultQueryTimeout
     }
 
     /** {@inheritDoc} */
-    @Override protected void executeQuery(String sql, long timeout) throws Exception {
+    @Override protected void executeQuery(String sql, int timeout) throws Exception {
         try (IgniteClient cli = G.startClient(new ClientConfiguration().setAddresses("127.0.0.1"))) {
-            cli.query(new SqlFieldsQuery(sql).setTimeout((int)timeout, TimeUnit.MILLISECONDS)).getAll();
+            cli.query(new SqlFieldsQuery(sql).setTimeout(timeout, TimeUnit.MILLISECONDS)).getAll();
         }
     }
